+++ lib/godmin/engine_wrapper.rb
@@ -17,7 +17,7 @@ module Godmin
     def namespaced_path
       @namespaced_path ||= begin
         if namespaced?
-          namespace.name.classify.split("::").map(&:underscore)
+          namespace.name.split("::").map(&:underscore)
         else
           []
         end